backend/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pillow
numpy
opencv-python
reportlab
grad-cam
python-multipart
